fix int64

+++ src/define/store_engine.go
@@ -3,13 +3,13 @@ package define
 import "log"
 
 type Engine interface {
-    GetUrl(id int) (string, error)
+    GetUrl(id int64) (string, error)
     StoreUrl(url string) (int64, error)
 }
 
 type MysqlEngine struct {}
 
-func (my *MysqlEngine) GetUrl(id int) (string, error) {
+func (my *MysqlEngine) GetUrl(id int64) (string, error) {
     var url []byte
     row := Db.QueryRow("select origin_url from short where id = ?", id)
     err := row.Scan(&url)
@@ -38,7 +38,7 @@ func (my *MysqlEngine) StoreUrl(url string) (int64, error) {
     return lastInsertId, nil
 }
 
-func (my *MysqlEngine) updateModifyTime(id int) {
+func (my *MysqlEngine) updateModifyTime(id int64) {
     if _, err := Db.Exec("update short set last_modify_time = now() where id = ?", id); err != nil {
         log.Println("update modify time failed")
     }
